refactor(cli): reorganize CLI code into gptme/cli/ module
Move CLI-related code into a dedicated gptme/cli/ module to clearly
separate the interface/UX from the core logic.

Reorganization:
- gptme/cli.py → gptme/cli/main.py (main CLI entry point)
- gptme/doctor.py → gptme/cli/doctor.py (diagnostics command)
- gptme/onboard.py → gptme/cli/onboard.py (onboarding wizard)
- gptme/setup.py → gptme/cli/setup.py (setup utilities)
- gptme/util/cli.py → gptme/cli/util.py (utility commands)

The main entry point 'gptme' continues to work via gptme/cli/__init__.py.

Updated entry points in pyproject.toml:
- gptme-util → gptme.cli.util:main
- gptme-doctor → gptme.cli.doctor:main
- gptme-onboard → gptme.cli.onboard:main

Closes #1171
