+++ pyproject.toml
@@ -20,10 +20,10 @@ gptme = "gptme.cli:main"
 gptme-server = "gptme.server.cli:main"
 gptme-acp = "gptme.acp.__main__:main"
 gptme-eval = "gptme.eval.main:main"
-gptme-util = "gptme.util.cli:main"
+gptme-util = "gptme.cli.util:main"
 gptme-agent = "gptme.agent.cli:main"
-gptme-doctor = "gptme.doctor:main"
-gptme-onboard = "gptme.onboard:main"
+gptme-doctor = "gptme.cli.doctor:main"
+gptme-onboard = "gptme.cli.onboard:main"
 gptme-wut = "gptme.wut:main"
 gptme-dspy = "gptme.eval.dspy:main"
